Refactored WebLoader to load in NodeJS
*	This fixes #23 by implementing a NodeJS Loader for JPAK / JMS
*	The WebLoader now is called Loader because it is universal
