+++ .gitignore
@@ -1,2 +1,3 @@
 node_modules
 3rdparty
+**/*.pyc
